src/requirements.txt
amazon.ion
aws-kinesis-agg
requests_aws4auth
elasticsearch